connect database to backend
